+++ android/src/main/kotlin/flutter/curiosity/CuriosityPlugin.kt
@@ -2,7 +2,6 @@ package flutter.curiosity
 
 import android.content.Context
 import android.content.Intent
-import android.graphics.BitmapFactory
 import android.graphics.Rect
 import android.util.Log
 import android.view.ViewTreeObserver
@@ -78,20 +77,14 @@ class CuriosityPlugin : ActivityAware, FlutterPlugin, MethodChannel.MethodCallHa
                 Tools.getGPSStatus(activityBinding.activity)
             )
 
-            "saveImageToGallery" -> {
-                val bytes = call.argument<ByteArray>("bytes")!!
-                val quality = call.argument<Int>("quality")!!
-                val name = call.argument<String?>("name")
-                val bitmap = BitmapFactory.decodeByteArray(bytes, 0, bytes.size)
-                result.success(
-                    ImageGalleryTools.saveImageToGallery(context, bitmap, quality, name)
-                )
+            "saveBytesImageToGallery" -> {
+                result.success(ImageGalleryTools.saveBytesImage(context, call))
             }
 
-            "saveFileToGallery" -> {
+            "saveFilePathToGallery" -> {
                 val filePath = call.argument<String>("filePath")!!
                 val name = call.argument<String?>("name")
-                result.success(ImageGalleryTools.saveFileToGallery(context, filePath, name))
+                result.success(ImageGalleryTools.saveFilePath(context, filePath, name))
             }
 
             else -> result.notImplemented()